Add tailwind
- Change page title
- Cleanup tsconfig
- Move hiber3d specific js into hiber3d folder
