+++ frontend/package.json
@@ -12,20 +12,24 @@
   "dependencies": {
     "@hiber3d/interop": "file:../game/build",
     "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-    "vite-multiple-assets": "^2.2.4"
+    "react-dom": "^18.3.1"
   },
   "devDependencies": {
     "@eslint/js": "^9.13.0",
     "@types/react": "^18.3.12",
     "@types/react-dom": "^18.3.1",
     "@vitejs/plugin-react": "^4.3.3",
+    "autoprefixer": "^10.4.20",
     "eslint": "^9.13.0",
     "eslint-plugin-react-hooks": "^5.0.0",
     "eslint-plugin-react-refresh": "^0.4.14",
     "globals": "^15.11.0",
+    "postcss": "^8.4.49",
+    "tailwindcss": "^3.4.16",
     "typescript": "~5.6.2",
     "typescript-eslint": "^8.11.0",
-    "vite": "^5.4.10"
+    "vite": "^5.4.10",
+    "vite-multiple-assets": "^2.2.4",
+    "vite-tsconfig-paths": "^5.1.3"
   }
 }
