+++ frontend/src/App.tsx
@@ -1,7 +1,9 @@
-import { Hiber3D } from "./components/Hiber3D";
+import { Hiber3D } from "hiber3d/components/Hiber3D";
 
 export const App = () => (
   <Hiber3D>
-    <h1>Hello Game World!</h1>
+    <div className="relative w-full h-dvh flex items-center justify-center">
+      <h1 className="text-white">Hello Game!</h1>
+    </div>
   </Hiber3D>
 );
